+++ bpdm-orchestrator/src/main/kotlin/org/eclipse/tractusx/bpdm/orchestrator/controller/v6/GoldenRecordTaskLegacyServiceMapper.kt
@@ -150,6 +150,7 @@ class GoldenRecordTaskLegacyServiceMapper(
                 legalEntityHeaderScriptVariants = mutableListOf(),
                 siteHeaderScriptVariants = mutableListOf(),
                 legalEntityGoldenRecordRelations = mutableListOf(),
+                siteGoldenRecordRelations = mutableListOf(),
                 addressGoldenRecordRelations = mutableListOf()
             )
         }