+++ bpdm-orchestrator/src/main/kotlin/org/eclipse/tractusx/bpdm/orchestrator/service/ResponseMapper.kt
@@ -110,6 +110,7 @@ class ResponseMapper {
                     hasChanged = siteHasChanged,
                     siteMainAddress = toPostalAddress(businessPartner, PostalAddressDb.Scope.SiteMainAddress),
                     scriptVariants = toSiteScriptVariants(businessPartner),
+                    goldenRecordRelations = toSiteGoldenRecordRelations(businessPartner),
                     updatedAt = siteUpdatedAt?.instant
                 )
             }
@@ -319,6 +320,10 @@ class ResponseMapper {
     fun toLegalEntityGoldenRecordRelations(businessPartner: GoldenRecordTaskDb.BusinessPartner) =
         businessPartner.legalEntityGoldenRecordRelations
             .map { LegalEntityGoldenRecordRelation(it.relationType, it.sourceBpn, it.targetBpn) }
+
+    fun toSiteGoldenRecordRelations(businessPartner: GoldenRecordTaskDb.BusinessPartner) =
+        businessPartner.siteGoldenRecordRelations
+            .map { SiteGoldenRecordRelation(it.relationType, it.sourceBpn, it.targetBpn) }
 
     fun toAddressGoldenRecordRelations(businessPartner: GoldenRecordTaskDb.BusinessPartner, scope: AddressGoldenRecordRelationDb.Scope) =
         businessPartner.addressGoldenRecordRelations